simple produce/consume complete, example updated
produce() and consume() now *work*.  produce is pretty
straightforward, and doesn't do anything fancy (make sure your
messages are in the right format!)

Consume() right now is a bit of a hack - it only returns a
reference to the buffer location of the message.  You will need
to copy that memory out before doing anything else, or it will
be overwritten.